+++ src/main/java/persistence/sql/dml/InsertDMLQueryBuilder.java
@@ -1,12 +1,22 @@
 package persistence.sql.dml;
 
 import persistence.sql.dbms.DbmsStrategy;
+import persistence.sql.entitymetadata.model.EntityColumn;
 import persistence.sql.entitymetadata.model.EntityColumns;
 import persistence.sql.entitymetadata.model.EntityTable;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
 public class InsertDMLQueryBuilder<E> extends DMLQueryBuilder<E> {
+
+    private E entity;
+    private List<String> defaultValuesClause;
+
     public InsertDMLQueryBuilder(DbmsStrategy dbmsStrategy, E entity) {
-        super(dbmsStrategy, entity);
+        super(dbmsStrategy, (Class<E>) entity.getClass());
+        this.entity = entity;
+        this.defaultValuesClause = initDefaultValuesClause();
     }
 
     @Override
@@ -17,9 +27,22 @@ public class InsertDMLQueryBuilder<E> extends DMLQueryBuilder<E> {
                 createInsertValueClause());
     }
 
-    /**
-     * insert into table (column1, column2, column3) values (value1, value2, value3)
-     */
+    private List<String> initDefaultValuesClause() {
+        return entityTable.getColumns()
+                .stream()
+                .map(this::getEntityColumnValue)
+                .map(String::valueOf)
+                .collect(Collectors.toList());
+    }
+
+    private <V> V getEntityColumnValue(EntityColumn<E, V> entityColumn) {
+        if (entityColumn.getType() == String.class) {
+            return (V) ("'" + entityColumn.getValue(entity) + "'");
+        }
+
+        return entityColumn.getValue(entity);
+    }
+
     private String createInsertColumnsClause() {
        return "(" + String.join(", ", defaultColumnsClause) + ")";
     }
